split up integration tests so they are easier to extend and track
